requirements.txt
acvl_utils==0.2.5
batchgenerators==0.25.1
batchgeneratorsv2==0.2.3
blosc2==3.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
connected-components-3d==3.23.0
contourpy==1.3.1
cycler==0.12.1
dicom2nifti==2.5.1
dynamic_network_architectures==0.3.1
einops==0.8.1
et_xmlfile==2.0.0
fft-conv-pytorch==1.2.0
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
future==1.0.0
graphviz==0.20.3
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
natsort==8.4.0
ndindex==1.9.2
networkx==3.3
nibabel==5.3.2
nnunetv2==2.6.0
numexpr==2.10.2
numpy==2.1.2
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
py-cpuinfo==9.0.0
pydicom==3.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==70.2.0
SimpleITK==2.4.1
six==1.17.0
slideio==2.7.1
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.13
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
traceback2==1.4.0
typing_extensions==4.12.2
tzdata==2025.2
unittest2==1.1.0
urllib3==2.3.0
yacs==0.1.8